Add current user request in api
