+++ api/src/routes/auth.route.ts
@@ -5,3 +5,5 @@ export const authRoute = Router();
 
 authRoute.post('/signUp', AuthController.signUp);
 authRoute.post('/signIn', AuthController.signIn);
+authRoute.post('/signOut', AuthController.signOut);
+authRoute.get('/currentUser', AuthController.currentUser);
