--- client/src/pages/signOut/index.tsx
@@ -1,8 +1,0 @@
-
-const SignOut = () => {
-  return (
-    <div>SignOut</div>
-  )
-}
-
-export default SignOut;
